+++ src/test/groovy/fr/nereide/test/reference/GenericRefTestCase.groovy
@@ -17,11 +17,12 @@
 
 package fr.nereide.test.reference
 
-import com.intellij.lang.properties.references.PropertyReference
 import com.intellij.psi.PsiReference
 import com.intellij.psi.impl.source.resolve.reference.impl.PsiMultiReference
 import fr.nereide.test.GenericOfbizPluginTestCase
+import org.junit.Ignore
 
+@Ignore('Setup class, No tests here')
 class GenericRefTestCase extends GenericOfbizPluginTestCase {
 
     @Override
@@ -40,17 +41,31 @@ class GenericRefTestCase extends GenericOfbizPluginTestCase {
         myFixture.configureByFile(file)
         PsiReference ref = myFixture.getReferenceAtCaretPositionWithAssertion()
         if (ref instanceof PsiMultiReference) {
+            PsiMultiReference multiRef = ref
             ref = ref.getReferences().find { expectedRefType.isAssignableFrom(it.getClass()) }
+            assertNoOtherRefType(multiRef, expectedRefType)
         }
         assert expectedRefType.isAssignableFrom(ref.getClass())
         if (strict) {
             assertEquals expectedRefValueName, ref.getElement().getName() ?
-                    ref.getElement().getName() : getSafeTextInReference(ref)
-                    as String
+                    ref.getElement().getName() : getSafeTextInReference(ref) as String
         } else {
             assert ref.getElement().getText().contains(expectedRefValueName)
         }
         assertNotNull "Reference for $expectedRefValueName not found", ref.resolve()
+    }
+
+    /**
+     * Checks that no other reference type was found for an element.
+     * For example, no screen reference resolved from a form
+     * @param multiRef
+     * @param expectedRefType
+     */
+    private static void assertNoOtherRefType(PsiMultiReference multiRef, Class expectedRefType) {
+        assert (multiRef.getReferences() as List).stream()
+                .filter { !expectedRefType.isAssignableFrom(it.getClass()) }
+                .filter { it.getClass().getName().contains('fr.nereide') }
+                .toList().size() == 0
     }
 
     static String getSafeTextInReference(PsiReference ref) {
@@ -63,10 +78,4 @@ class GenericRefTestCase extends GenericOfbizPluginTestCase {
         configureByFileAndTestRefTypeAndValue(file, expectedRefType, expectedRefValueName, true)
     }
 
-    /**
-     * Temporary workaround for tests to stay green
-     */
-    void testDummy() {
-        assert true
-    }
 }