+++ src/test/groovy/fr/nereide/test/reference/TestReferenceInXmlCompound.groovy
@@ -1,6 +1,7 @@
 package fr.nereide.test.reference
 
 import com.intellij.psi.impl.source.resolve.reference.impl.providers.FileReference
+import fr.nereide.reference.xml.GridReference
 import fr.nereide.reference.xml.JavaMethodReference
 import fr.nereide.reference.xml.RequestMapReference
 import fr.nereide.reference.common.ServiceReference
@@ -19,13 +20,14 @@ class TestReferenceInXmlCompound extends GenericRefTestCase {
         return "src/test/resources/testData/compound"
     }
 
-    private void configureAndMoveFileAndTestRefTypeAndValue(String file, Class expectedRefType, String expectedRefValueName) {
-        configureAndMoveFileAndTestRefTypeAndValue(file, expectedRefType, expectedRefValueName, true)
+    protected void doTest(Class expectedClass, String expectedName) {
+        doTest(expectedClass, expectedName, true)
     }
 
-    private void configureAndMoveFileAndTestRefTypeAndValue(String file, Class expectedRefType, String expectedRefValueName, boolean strict) {
+    protected void doTest(Class expectedClass, String expectedName, boolean strict) {
+        String file = "${this.getTestName(false).replaceAll('test', '')}.xml"
         myFixture.moveFile(file, MOVE_TO)
-        configureByFileAndTestRefTypeAndValue(file, expectedRefType, expectedRefValueName, strict)
+        configureByFileAndTestRefTypeAndValue(file, expectedClass, expectedName, strict)
     }
 
     //=====================================
@@ -33,104 +35,87 @@ class TestReferenceInXmlCompound extends GenericRefTestCase {
     //=====================================
 
     void testCpdFormReferenceFromCpdScreen() {
-        String file = 'CpdFormReferenceFromCpdScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, FormReference.class, 'FooForm')
+        doTest(FormReference.class, 'FooForm')
     }
 
     void testCpdGridReferenceFromCpdScreen() {
-        String file = 'CpdGridReferenceFromCpdScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, FormReference.class, 'FooGrid')
+        doTest(GridReference.class, 'MyCompoundElement')
     }
 
     void testCpdScreenReferenceFromCpdScreen() {
-        String file = 'CpdScreenReferenceFromCpdScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, ScreenReference.class, 'BarScreen')
+        doTest(ScreenReference.class, 'BarScreen')
     }
 
     void testCpdScreenDecoratorReferenceFromCpdScreen() {
-        String file = 'CpdScreenDecoratorReferenceFromCpdScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, ScreenReference.class, 'FooDecoratorScreen')
+        doTest(ScreenReference.class, 'FooDecoratorScreen')
     }
 
     void testExternalFormRefFromCpdScreen() {
-        String file = 'ExternalFormRefFromCpdScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, FormReference.class, 'OneFormAmongOthers')
+        doTest(FormReference.class, 'OneFormAmongOthers')
     }
 
     void testExternalScreenRefFromCpdScreen() {
-        String file = 'ExternalScreenRefFromCpdScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, ScreenReference.class, 'MundaneScreen')
+        doTest(ScreenReference.class, 'MundaneScreen')
     }
 
     void testCpdMenuRefFromCpdScreen() {
-        String file = 'CpdMenuRefFromCpdScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, MenuReference.class, 'FooMenu')
+        doTest(MenuReference.class, 'FooMenu')
     }
 
     void testExternalMenuRefFromCpdScreen() {
-        String file = 'ExternalMenuRefFromCpdScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, MenuReference.class, 'BarMenu')
+        doTest(MenuReference.class, 'BarMenu')
     }
 
     void testEntityRefFromCpdScreen() {
-        String file = 'EntityRefFromCpdScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, EntityReference.class, 'Entito')
+        doTest(EntityReference.class, 'Entito')
     }
 
     void testFileRefFromCpdScreen() {
-        String file = 'FileRefFromCpdScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, FileReference.class, 'footemplate', false)
+        doTest(FileReference.class, 'footemplate', false)
     }
 
     void testCpdScreenRefFromExtScreen() {
-        String file = 'CpdScreenRefFromExtScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, ScreenReference.class, 'MyScreenInCpd', false)
+        doTest(ScreenReference.class, 'MyScreenInCpd', false)
     }
 
     void testCpdFormRefFromExtScreen() {
-        String file = 'CpdFormRefFromExtScreen.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, FormReference.class, 'MyFormInCpd', false)
+        doTest(FormReference.class, 'MyFormInCpd', false)
     }
+
     //=====================================
     //              FORMS TESTS
     //=====================================
 
     void testCpdRequestMapRefFromCpdForm() {
-        String file = 'CpdRequestMapRefFromCpdForm.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, RequestMapReference.class, 'MyFooRequest')
+        doTest(RequestMapReference.class, 'MyFooRequest')
     }
 
     void testCpdScreenRefFromCpdFieldForm() {
-        String file = 'CpdScreenRefFromCpdFieldForm.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, ScreenReference.class, 'MyPitaScreen', false)
+        doTest(ScreenReference.class, 'MyPitaScreen', false)
     }
 
     void testCpdFormRefFromCpdFieldForm() {
-        String file = 'CpdFormRefFromCpdFieldForm.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, FormReference.class, 'MyPitaForm')
+        doTest(FormReference.class, 'MyPitaForm')
     }
 
     void testCpdMenuRefFromCpdFieldForm() {
-        String file = 'CpdMenuRefFromCpdFieldForm.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, MenuReference.class, 'MyPitaMenu')
+        doTest(MenuReference.class, 'MyPitaMenu')
     }
+
     //=====================================
     //         REQUEST MAP TESTS
     //=====================================
 
     void testCpdViewMapRefFromCpdRequestMap() {
-        String file = 'CpdViewMapRefFromCpdRequestMap.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, ViewMapReference.class, 'myFooResponseInCpd')
+        doTest(ViewMapReference.class, 'myFooResponseInCpd')
     }
 
     void testServiceRefFromCpdRequestMap() {
-        String file = 'ServiceRefFromCpdRequestMap.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, ServiceReference.class, 'DonateToQuadratureDuNet')
+        doTest(ServiceReference.class, 'DonateToQuadratureDuNet')
     }
 
     void testJavaEventRefFromCpdRequestMap() {
-        String file = 'JavaEventRefFromCpdRequestMap.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, JavaMethodReference.class, 'login')
+        doTest(JavaMethodReference.class, 'login')
     }
 
     //=====================================
@@ -138,7 +123,6 @@ class TestReferenceInXmlCompound extends GenericRefTestCase {
     //=====================================
 
     void testCpdScreenRefFromCpdViewMap() {
-        String file = 'CpdScreenRefFromCpdViewMap.xml'
-        configureAndMoveFileAndTestRefTypeAndValue(file, ScreenReference.class, 'MyFooScreenInCpd', false)
+        doTest(ScreenReference.class, 'MyFooScreenInCpd', false)
     }
 }